+++ libs/client-sdk/tests/client_c_bindings_test.cpp
@@ -27,7 +27,7 @@ TEST(ClientBindings, MinimalInstantiation) {
 
     char const* version = LDClientSDK_Version();
     ASSERT_TRUE(version);
-    ASSERT_STREQ(version, "3.6.3");  // {x-release-please-version}
+    ASSERT_STREQ(version, "3.8.1");  // {x-release-please-version}
 
     LDClientSDK_Free(sdk);
 }
@@ -148,6 +148,8 @@ TEST(ClientBindings, GetStatusOfOfflineClient) {
     bool success = false;
     LDClientSDK_Start(sdk, 3000, &success);
 
+    EXPECT_TRUE(success);
+
     LDDataSourceStatus status_2 = LDClientSDK_DataSourceStatus_Status(sdk);
     EXPECT_EQ(LD_DATASOURCESTATUS_STATE_OFFLINE,
               LDDataSourceStatus_GetState(status_2));
@@ -155,6 +157,8 @@ TEST(ClientBindings, GetStatusOfOfflineClient) {
     EXPECT_EQ(nullptr, LDDataSourceStatus_GetLastError(status_2));
 
     EXPECT_NE(0, LDDataSourceStatus_StateSince(status_2));
+
+    EXPECT_TRUE(LDClientSDK_Initialized(sdk));
 
     LDDataSourceStatus_Free(status_1);
     LDDataSourceStatus_Free(status_2);
@@ -193,4 +197,25 @@ TEST(ClientBindings, ComplexDataSourceStatus) {
     EXPECT_EQ(404, LDDataSourceStatus_ErrorInfo_StatusCode(info));
 
     LDDataSourceStatus_ErrorInfo_Free(info);
+}
+
+TEST(ClientBindings, TestDataSourceStatusStateName) {
+    ASSERT_STREQ(LDDataSourceStatus_State_Name(LD_DATASOURCESTATUS_STATE_VALID,
+                                               "unknown"),
+                 "VALID");
+    ASSERT_STREQ(LDDataSourceStatus_State_Name(
+                     LD_DATASOURCESTATUS_STATE_OFFLINE, "unknown"),
+                 "OFFLINE");
+    ASSERT_STREQ(LDDataSourceStatus_State_Name(
+                     LD_DATASOURCESTATUS_STATE_INITIALIZING, "unknown"),
+                 "INITIALIZING");
+    ASSERT_STREQ(LDDataSourceStatus_State_Name(
+                     LD_DATASOURCESTATUS_STATE_SHUTDOWN, "unknown"),
+                 "SHUTDOWN");
+    ASSERT_STREQ(LDDataSourceStatus_State_Name(
+                     LD_DATASOURCESTATUS_STATE_INTERRUPTED, "unknown"),
+                 "INTERRUPTED");
+    ASSERT_STREQ(LDDataSourceStatus_State_Name(
+                     LD_DATASOURCESTATUS_STATE_UNUSED_MAXVALUE, "unknown"),
+                 "unknown");
 }