+++ libs/server-sdk/src/data_systems/background_sync/sources/polling/polling_data_source.cpp
@@ -4,11 +4,13 @@
 #include <launchdarkly/network/http_error_messages.hpp>
 
 #include <launchdarkly/serialization/json_flag.hpp>
-#include <launchdarkly/serialization/json_primitives.hpp>
+#include <launchdarkly/detail/serialization/json_primitives.hpp>
 #include <launchdarkly/serialization/json_sdk_data_set.hpp>
 #include <launchdarkly/server_side/data_source_status.hpp>
 
 #include <launchdarkly/server_side/config/builders/all_builders.hpp>
+
+#include "../../detail/payload_filter_validation/payload_filter_validation.hpp"
 
 #include <boost/json.hpp>
 
@@ -20,13 +22,31 @@ static char const* const kErrorPutInvalid =
 static char const* const kCouldNotParseEndpoint =
     "Could not parse polling endpoint URL";
 
+static char const* const kInvalidFilterKey =
+    "Invalid payload filter configured on polling data source, full "
+    "environment "
+    "will be fetched.\nEnsure the filter key is not empty and was copied "
+    "correctly from LaunchDarkly settings";
+
 static network::HttpRequest MakeRequest(
+    Logger const& logger,
     config::built::BackgroundSyncConfig::PollingConfig const& polling_config,
     config::built::ServiceEndpoints const& endpoints,
     config::built::HttpProperties const& http_properties) {
     auto url = std::make_optional(endpoints.PollingBaseUrl());
 
     url = network::AppendUrl(url, polling_config.polling_get_path);
+
+    if (polling_config.filter_key && url) {
+        if (detail::ValidateFilterKey(*polling_config.filter_key)) {
+            url->append("?filter=" + *polling_config.filter_key);
+            LD_LOG(logger, LogLevel::kDebug)
+                << "using payload filter '" << *polling_config.filter_key
+                << "'";
+        } else {
+            LD_LOG(logger, LogLevel::kError) << kInvalidFilterKey;
+        }
+    }
 
     network::HttpRequest::BodyType body;
     network::HttpMethod method = network::HttpMethod::kGet;
@@ -54,7 +74,8 @@ PollingDataSource::PollingDataSource(
       status_manager_(status_manager),
       requester_(ioc, http_properties.Tls()),
       polling_interval_(data_source_config.poll_interval),
-      request_(MakeRequest(data_source_config, endpoints, http_properties)),
+      request_(
+          MakeRequest(logger_, data_source_config, endpoints, http_properties)),
       timer_(ioc),
       sink_(nullptr) {
     if (polling_interval_ < data_source_config.min_polling_interval) {